we don't use and never have used shared libraries.
On top of that fact dpkg-shlibdep was never called from rules so the
substitue file would never be generated.  And if if it was generated it
would be empty as we have no binaries in the build dir that use shared
libs.  So this just produced warnings in the build
  `dpkg-gencontrol: warning: Depends field of package jenkins: unknown
  substitution variable ${shlibs:Depends}`