Define the required "_gpg_name" macro.
The _gpg_name macro is required by rpmbuild but it is normally defined per
user in ~/.rpmmacros.  However on a CI sytem we don't want to have
something hard coded and left on the file system, and we also expicitly
state what keys to use this macro is completely ignored anyway.  So to
prevent build issues just define the macro to something.