[JENKINS-37721] the service in the jenkins.xml must match what wix installed
Make sure that the service name used by the jenkins.exe is the same as the
service name registered by the msi.

Not an issue for vanilla Jenkins as the registry is case insensitive and
the fact that jenkins != Jenkins is ok, however when the artifact name
contains a dash such as jenkins-oc and the camel artifact name is
forbidden to cotain dashes such that they artifact name is not jusy a
lower case equivallent of camelartifactname then things do not work.

Whilst I was there I also corrected the description of the service in the
XML - ot that it is used , but more because it looks better (and someone
could use this to install the service if they did not want to use the MSI
by copying the files and running `jenkins.exe install`