Fix postinstall script for OS X
The installer looks for an available uid bellow 500. The logic of
the script list the uids of the system in descending order and pickup
the first below 500 and adds 1 to it. This causes 2 issues:

1) If uid 499 exist it will try to use 500. 500 means in the script that
   all uids are used. This happens even though there is romm in lower
   uids.

2) The script does not check if the uid picked up is already in use
   which may cause some conflicts.

This fix walks the allocated uids below 500 takes every candidate and
checks whether the formed uid is already used.

More details can be found in:

https://issues.jenkins-ci.org/browse/JENKINS-43284
https://groups.google.com/forum/#!msg/jenkinsci-dev/ip-GAlRCetA/uYYKGksRBAAJ