Update Windows 'latest' symlink after uploading new msi
The 'latest' symlink allows the Download page on jenkins.io to have a
constant URL that points to http://mirrors.jenkins.io/windows/latest .
It relies on the production build process to update the symlink to point
to the most recent build.

Step 1 to fix https://issues.jenkins-ci.org/browse/WEBSITE-758 and
https://github.com/jenkins-infra/jenkins.io/issues/3476 .

Step 2 requires a change in the Downloads page on jenkins.io.