Convert RPMs from sysvinit to systemd
Since EL7 systemd is the default init system. This moves away from the
legacy sysvinit script to a native unit file.

It no longer ships the sysconfig file by default, but it is still read
for compatibility. Users are encouraged to use systemd drop in files
which are a more native solution.

The deprecated PreReq is replaced with a stricter Requires(pre). Ideally
sysusers would be used, but the packaging macros in EL7 are different
from more modern versions. For now this can be considered a TODO.