Fix javaHome arg in Debian defualt file comment
The `--javahome` argument that was listed there is rejected by
Jenkins when it is included in the JENKINS_ARGS variable.

The `--javaHome` argument is accepted by Jenkins but does not seem to
set the JAVA_HOME environment variable in Jenkins 2.330 on Debian 11.
The JAVA_HOME environment variable is not shown in Jenkins 2.330 in
the "Manage Jenkins" -> "System Information" page of Jenkins 2.330 and
is not shown in the environment variable list reported by a job run on
the Jenkins 2.330 controller.

The `--javaHome` argument is accepted by Jenkins and sets the
JAVA_HOME environment variable in Jenkins 2.333 when using pull
request #266 (systemd support) and is shown in the environment
variable list reported by a job run on the Jenkins 2.333 controller
installed with the systemd support provided in pull request #266.