Do not publish MSI package or update its pages
MSI installer needs to be signed with a code signing certificate.
Our code signing certificate has expired.  Attorneys from the Linux
Foundation and the Continuous Delivery Foundation are working to renew
the code signing certificate. Once that is completed, these changes can
be removed and we'll resume publishing the MSI installer for Windows.

Exit early from make.ps1 in case it is used in the build process.