Use libfontconfig1 for container consistency
The container images use the libfontconfig1 package instead of using the
fontconfig package.  The libfontconfig1 is a dependency of the fontconfig
package and seems like a good choice for consistency between container
images and versions installed with the operating system package manager.

https://github.com/search?q=repo%3Ajenkinsci%2Fdocker%20libfontconfig1&type=code
shows the current Debian containers use libfontconfig1 while the UBI
containers and the AlmaLinux containers use fontconfig.