Remove Java 11 from MSI installer (Windows)
Jenkins weekly 2.463 requires Java 17 or Java 21.  Do not offer Java 11
to Windows users as a valid Java installation for Jenkins.

Testing done

Built the Windows installer with these changes and confirmed that Java 8
and Java 11 are both rejected as invalid Java versions and that Java 17
and Java 21 are accepted as valid Java versions.