Fix incorrect conditional in Java 17 version check
Conditional was checking that the JAVA_EXE_VERSION = "17" when it should
have been checking JAVA_EXE_VERSION <> "17" like the line before it and
the line after it.

Also fixed the inconsistent name of the property.  All the other
properties related to JAVA_HOME are prefixed with JavaHome.  This one
was prefixed with JavaJome.

Testing done:

Confirmed that the expected error message is displayed when I choose a
Java directory with Java 8.

Confirmed that the expected error message is displayed when I choose a
Java directory that does not contain Java.

Confirmed that no error message is displayed when I choose a Java
directory with Java 11 or Java 17 or Java 21.